Cleaned up files
Cleaned up files and initial commit of iteration 3 tasks
