+++ dog tinder/dog tinder/Database/Dog.swift
