+++ dog tinder/dog tinder/dog_tinderApp.swift
@@ -15,9 +15,3 @@ struct dog_tinderApp: App {
         }
     }
 }
-
-struct dog_tinderApp_Previews: PreviewProvider {
-    static var previews: some View {
-        /*@START_MENU_TOKEN@*/Text("Hello, World!")/*@END_MENU_TOKEN@*/
-    }
-}
